+++ tests/Ui/Bundle/AdminBundle/Controller/Happening/UpdateActionTest.php
@@ -160,6 +160,7 @@ class UpdateActionTest extends TestCase
         $this->happening->isSidebarAllowed()->willReturn(true);
         $this->happening->isWebinarRecorded()->shouldBeCalled()->willReturn(true);
         $this->happening->allowWebinarOnHLS()->willReturn(true);
+        $this->happening->isWebinarRecordSentToSpeakers()->willReturn(true);
 
         $form = $this->prophesize(Form::class);
         $formView = $this->prophesize(FormView::class);
@@ -242,6 +243,7 @@ class UpdateActionTest extends TestCase
         $this->happening->isSidebarAllowed()->willReturn(true);
         $this->happening->isWebinarRecorded()->shouldBeCalled()->willReturn(true);
         $this->happening->allowWebinarOnHLS()->willReturn(true);
+        $this->happening->isWebinarRecordSentToSpeakers()->willReturn(true);
 
         $form = $this->prophesize(Form::class);
         $allowHlsFormField = $this->prophesize(Form::class);